InAppNotificationOverlay: split banner views into peer files
InAppNotificationOverlay.swift was 728 lines holding three independent
top-level views and a preview host. Splits:

- InAppNotificationBanner (the expanded/collapsed main banner with drag
  gestures and previews) → InAppNotificationBanner.swift, taking the
  PreviewHost helper and all 10 banner-flavor #Preview blocks with it.
- FastTrackBanner (the compact "Copied to Clipboard"-style banner) →
  FastTrackBanner.swift with its single preview.

The overlay file itself now stays at 300 lines and only holds the queue
manager that orchestrates which banner to show.

Both new files are added to the ShellbeeWidgetsExtension membership
exception list.

Refs #33
